Minimum & Maximum support and basic datetime default value support
Added support for for setting minimum and maximum values for numeric values, and added support for setting date with a default value (previously the entire expression would be converted to a string without being evaluated which would throw a bug when importing into ThingWorx). I have only added support for setting date with a single string argument.